requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.3.0
autopep8==1.6.0
backcall==0.2.0
blis==0.7.5
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cycler==0.11.0
cymem==2.0.6
decorator==5.1.0
filelock==3.4.2
fonttools==4.28.5
frozenlist==1.2.0
fsspec==2021.11.1
future==0.18.2
gensim==4.1.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.43.0
huggingface-hub==0.2.1
idna==3.3
importlib-metadata==4.10.0
ipython==7.30.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
langcodes==3.3.0
lr-scheduler @ git+https://github.com/sooftware/pytorch-lr-scheduler@c54f66ad9276ada40551351b67e395b9415d5e26
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
multidict==5.2.0
murmurhash==1.0.6
nltk==3.6.7
numpy==1.21.5
oauthlib==3.1.1
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
preshed==3.0.6
prompt-toolkit==3.0.24
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydantic==1.8.2
pyDeprecate==0.3.1
Pygments==2.10.0
pyparsing==3.0.6
pyreadline==2.1
python-dateutil==2.8.2
pytorch-lightning==1.5.7
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.46
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.1.0
sentencepiece==0.1.96
six==1.16.0
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thinc==8.0.13
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
torch==1.10.1
torchmetrics==0.6.2
torchvision==0.11.2
tqdm==4.62.3
traitlets==5.1.1
transformers==4.15.0
treelib==1.6.1
typer==0.4.0
typing_extensions==4.0.1
urllib3==1.26.7
wasabi==0.9.0
wcwidth==0.2.5
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.6.0
